requirements.txt
langchain
python_dotenv
langchain_community
langchain_core
streamlit
## ollama pull orca-mini